+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 ### Added
+- `AnyVec::clone_empty_in`
 - `Stack` Mem/Allocator. 
 - `Mem` ala Allocator.
 - `reserve`